+++ packages/core/src/option/OptionControlProps.ts
@@ -7,8 +7,10 @@ export default interface OptionControlProps<V> {
   disabled?: boolean
   onClick: (value: V) => void
   onMouseEnter: (value: V) => void
+  onMouseLeave: React.MouseEventHandler
   children: RenderChild<{
     onClick: React.MouseEventHandler
     onMouseEnter: React.MouseEventHandler
+    onMouseLeave: React.MouseEventHandler
   }>
 }