+++ packages/desktop/src/header-menu/index.ts
@@ -1,0 +1,2 @@
+export * from './HeaderMenuItem'
+export * from './HeaderMenu'