+++ packages/mobile/src/checkbox/CheckboxProps.ts
@@ -1,0 +1,9 @@
+export default interface CheckboxProps {
+  disabled?: boolean
+  checked: boolean
+  label: string
+  tabIndex?: number
+  description?: string
+  autoFocus?: boolean
+  onChange: (value: boolean) => void
+}