+++ packages/mobile/src/typography/Caption.tsx
@@ -1,9 +1,11 @@
 import React, {FunctionComponent} from 'react'
 
-import {Typo} from '@qiwi/pijma-core'
+import {Breaker, Box, Stub, Typo, TypoProps} from '@qiwi/pijma-core'
 
 export interface CaptionProps {
   color?: 'default' | 'support' | 'inverse'
+  align?: TypoProps['align']
+  stub?: boolean
 }
 
 const CaptionColor: { [color in NonNullable<CaptionProps['color']>]: string } = {
@@ -12,18 +14,35 @@ const CaptionColor: { [color in NonNullable<CaptionProps['color']>]: string } = 
   inverse: '#fff',
 }
 
-export const Caption: FunctionComponent<CaptionProps> = ({color = 'support', children}) => (
-  <Typo
-    as="h6"
-    display="block"
-    size={3.5}
-    height={4}
-    weight={500}
-    color={CaptionColor[color]}
-    transform="uppercase"
-    spacing={1.5}
-    children={children}
-  />
+export const Caption: FunctionComponent<CaptionProps> = ({color = 'support', align, stub, children}) => (
+  stub ? (
+    <Box
+      ml={align === 'center' || align === 'right' ? 'auto' : undefined}
+      mr={align === 'center' ? 'auto' : undefined}
+      width={50}
+      maxWidth={1}
+    >
+      <Stub
+        top={1}
+        bottom={1}
+        width={50}
+        height={3}
+      />
+    </Box>
+  ) : (
+    <Typo
+      as="h6"
+      display="block"
+      size={3.5}
+      height={5}
+      weight={500}
+      color={CaptionColor[color]}
+      transform="uppercase"
+      spacing={1.5}
+      align={align}
+      children={<Breaker children={children}/>}
+    />
+  )
 )
 
 Caption.defaultProps = {