+++ packages/media/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@qiwi/pijma-media-v1.60.1](https://github.com/qiwi/pijma/compare/v1.60.0...v1.60.1) (2019-07-10)
+
+
+### Performance Improvements
+
+* up deps ([a7bcce6](https://github.com/qiwi/pijma/commit/a7bcce6))
+
 # [@qiwi/pijma-media-v1.38.0](https://github.com/qiwi/pijma/compare/v1.37.0...v1.38.0) (2019-04-16)